src/requirements.txt
fastapi
uvicorn
streamlit
faiss-cpu
sentence-transformers
numpy
requests
ollama
python-multipart
pydantic
